Converted Nodes & Labels to use absolute coordinates, preserving drag behavior.
This eliminates a lot of unnecessary 'translate' operations in the default
rendering and makes the dragged nodes clearer in the SVG output.

This required handling drag-moves differently - now labels have to be
translated just like the nodes. (But: now that this is done, the labels
are free to move to their own independent layer. Soon.)

Also:
* The SVG groups of nodes and flows now have helpful IDs on them.
* Some .attr and .style commands now use the grouped form. (Less noisy.)