Tweaked default diagram appearance
The 'Wages' node will now be the custom-color example. This makes it
easier to see the specific effect of the color code, which is to change
the node (and not necessarily the flow).

UI header text is slightly larger now.

Default node width and spacing are both slightly larger now.