Node moves are now preserved as long as possible! Fixes #27.
Added a persistent Node-move memory so SankeyMATIC will always keep track of
which ones have been moved to custom positions.

The moves are quantified in percentage terms, not pixel counts, so even when
the size of the diagram is changed, the overall rearrangement is preserved
fairly well.

Rearranging the order of your inputs has no effect on the memory -- moves are
remembered as long as that Node's name is present in the inputs. When that
name is changed or deleted, that's when the memory of its custom position goes
away.

--

Changes to the visible UI:

There's a new prominent mention of dragging just above the diagram (since it
is no longer a source of unpleasant surprises).

There is also a new 'Reset all moved Nodes' button if you just want to start
over.

The 'Advanced' UI panel has been disbanded:
* The Drag mention is now above the diagram.
* The Reverse-the-Graph checkbox and the Scale figure are now in the 'Layout'
panel.

One more nifty thing: If you Reverse the graph to be right-to-left, any moved
Nodes will correctly be placed at the mirror-image of their original
customized positions.