Updated the default Budget diagram to better demonstrate custom colors.
Ever since I added the 'outside-in' Flow color pattern, the example code for
customizing a Node's color has *also* been affecting a Flow's color. This is
potentially confusing.

Now the custom-Node-color example line sets the color of "Budget" instead of
"Wages". This means that the line is visibly controlling only a single Node's
color and no Flows.

After this change, the effect of clicking the color-rotation buttons makes a
bit more intuitive sense: *every* Flow's color is now going to cycle through
the set (except the one customized Flow color).