Fixed a minor issue with the 'Reset all moved Nodes' button.
Now, the button will once again be disabled/reset when you 1) move one or more
Nodes; 2) restore their positions with double-clicks.

(The button was staying active when there was nothing for it to do.)

This was caused by a silly coding error - you can't check for equality between
arrays with a simple '=='.