Bug fix for an issue with the Node border appearance in PNGs.
Canvg does not appear to support the "paint-order" attribute in the way that
it was used in the last commit, so PNG exports were not matching the SVG
as it appeared on the main page.

Switching to a two-<rect> solution for now - first draw a rect with just the
stroke, then draw a second rect on top of that with just the fill.
(Essentially this does manually what the "paint-order" attribute should do.)