We now handle unwanted "zero-width spaces" in pasted inputs. Fixes #54.
Some apps include the character U+200B when you copy text from them. This
interferes with the parsing routine, which has been treating the invisible
character as part of a node name.

SankeyMATIC will now treat that character as ignore-able whitespace when it
appears at the start or end of a line.

This will save some steps when pasting text from such a source.