Fixed occasional incorrect highlighting of flows in PNG/SVG output.
When you stop dragging a node, sometimes the pointer is also over a flow,
which triggers an increase in that flow's opacity and highlights the
labels at both ends.

These presentation changes are only intended to appear in the live diagram,
not in rendered output.

So now, when a node drag ends, SankeyMATIC explicitly looks for any
lingering flows which are still highlighted and turns off those effects
before rendering the PNG and SVG.

As part of this change, flows now have their own unique DOM ids.