Added a system for measuring labels, producing better highlight sizes.
Instead of having a hardcoded count of pixels for setting a label's
padding, we now base padding values on the current font's actual em-width
and (approximated) x-height.

As part of this, we also adjust for differences in how Firefox reports the
size of SVG text - the default highlights looked shorter in FF than they
did everywhere else.

The fun part of this feature is that all the text gets pre-rendered in a
new hidden SVG area placed just under the diagram. This way we can know how
big things will be before placing them in the main diagram.