Margins are now automatically adjusted when labels are on the outside!
In the past, when labels have been placed to the left or right of nodes,
you had to adjust the margins yourself to make them fit on the canvas.

This is a highly tedious process in the UI, so by default, labels were all
placed inside the diagram area to avoid it. However, that's a bad place to
put the labels for multiple reasons: it can be confusing to figure out
which node a label goes with; labels will collide/overlap more frequently;
etc. etc.

Now, that annoying adjustment step is *gone*. SankeyMATIC automatically
adjusts the space the diagram is given to ensure that any labels on the
ends will fit.

(In the edge case where there just is not enough room for both the flows
and the labels -- say, due to a very large label font size -- the diagram
wins & the labels will be partially cut off.)

Also:
* Removed the note saying you'll have to adjust the margins manually.