Converted the remaining numeric inputs to have HTML type 'number'.
In a desktop browser, this provides +/- stepper buttons to make
experimenting with new values easier.

In a mobile browser, it seems to have no effect today; they are still
treated as 'text' type inputs.

Converted the 'Node Border' input back to a numeric input instead of a
slider; now if you want a border > 9, it's possible to do that.