Refined the flow sorting algorithm some more, fixing some edge cases.
Estimating an inner point on a node where incoming flows should attach
themselves (and measure their slopes against) seemed like a good idea,
but it was not working for some data sets with very uneven flow sizes.

(The Camouflage spelling diagram from 2017 is one which demonstrated
the breakage, with a 1:52:1 relation between the first 3 flows.)

So now I’ve rewritten the function to just use the same basic visual
standard a person might use:
* Measure the slope from the top edge of a flow to the top edge of the
  available space (and from bottom edge to bottom edge).
* Whichever flow has the least ideal slope in that department should
  get the available spot at that end of the range.
* In the next pass, measure the remaining slopes anew in relation to
  the remaining space.

This change made it possible to simplify away several pieces of logic,
to the point where the total size of placeFlowsInNode() has dropped
by 25% (even accounting for all the comments).