Layout change: immediately-neighboring nodes are now centered as a group
We already make sure nodes are spaced far enough apart, but that's not
really enough.

This change adds a followup step to make sure nodes which have been
bumped out of their ideal positions are then re-centered *as a group*
in their best remaining position (relative to all of their in & out
flows).