The Node Spacing slider is now split into two: "Height" & "Spacing"
Up until now, the Node Spacing slider has really been controlling 2
variables simultaneously. In practice, it has meant:
1. Set the node height to be X%, and along with that,
2. Use the MAXIMUM SPACING POSSIBLE AT ALL TIMES.

Now you can control the spacing with an independent slider, which can
give you many more pleasant options for how exactly your diagram will
look.

As part of this change, the diagram is also continually re-centered in
the available space so that it appears stable (without this, the
diagram currently kind of creeps downwards as the layout is refined).