The Flow sorting algorithm can now place flows at the bottom of a Node
When the flows out of a Node do not add up to the Node's total, they
have always been attached to the top of that Node.

This doesn't look good when the flows lead to positions below that
Node.

The layout algorithm now checks for the case where those flows might
look better attached to the Node's bottom, and uses the bottom portion
of the Node in that case.