Layout algorithm change: When flows span stages, they get shadow nodes
In order to make flows which span stages preserve an appropriate amount
of room for all the other nodes in those same stages, SankeyMATIC now
creates ‘shadow’ nodes which occupy space in every stage that a flow
crosses.

As part of this change, there is a new checkbox under Layout >
Debugging: "Reveal Shadow Nodes". This makes it possible to see the
invisible layout elements we're working with.

A couple of side effects to be aware of:

1. The order of your inputs now matters even more than before; it
affects the positions of all the shadow nodes as well as the main
visible nodes. Checking "use the exact input order" can create some
pretty dramatic bends if the rows are not in a good order (but this
also means that you can control the flow arrangement more precisely
when you want).

2. For any diagram which only has one visible node per stage, you can
now adjust the spacing between those flows since the ‘solo’ nodes now
have something to push against. (Before this, moving the Node Spacing
slider had no effect at all on the layout of these diagrams.)