Overhauled how settings are validated internally. No visible changes.
The validation of the various settings is now data-driven, based on a
complete list in constants.js which contains types, validation criteria
and default values.

This allows a lot of custom lists & code in the process_sankey()
function to be deleted.

One side effect: the saved-diagram recipes now use the actual settings
names & values for radio buttons (e.g. node_theme: 'a') instead of
hacking our way around them by toggling a specific field (as in
theme_a_radio: 'true'). This is going to enable better syntax for the
importing of settings later.