Made the project easier to run from a simple git checkout
Updated the references to 3rd-party libraries D3 and canvg to load them
from a Content Delivery Network.

This should make checking out & running SankeyMATIC to be a much
smoother experience, since you won't have to install any other
libraries locally in order for it to run.

Resolves #59.

(The live-site version is likely to continue to load copies of these
scripts stored on the home server, but that will be handled in another
branch and won't impact the 'main' branch.)