One last rename of some fields before the import/export feature
As I was implementing the import/export of diagrams in text files, it
became clear that life would be simpler if every setting had the simpler
form 'prefix_specificname' instead of 'prefix_plus_multiple_words'.

Therefore, this change standardizes all settings to match that
pattern (in some cases leading to new prefixes).

Also:

* Node Width and Node Border settings are now constrained to only be as
  wide as the diagram; going over that value leads to rejection
  (similar to the Left and Right Margins).

* The numeric settings' default values are now stored as actual numbers
  in the data structure.