Feature: Settings can now be declared in the main input field
This means that nearly every aspect of a diagram can now be specified
in one multi-line string. (Manually-moved nodes will be supported
in an upcoming commit.)

Setting declarations have 3 parts: a category prefix, a specific
identifier, and a value. For text strings (like the prefix shown in
front of each number), the string should be single-quoted.
  * Examples: "size width 800", "labels color #222", "value prefix '$'"

Any time a setting is applied successfully, that line is then commented
out from the inputs area and a checkmark is added to it. This serves two
purposes:
  1) it confirms to the user that their directive was successful, and
  2) it prevents contention between the UI controls on the web page and
     the value in the user's source text.

Notes on accepted syntax variations:

* When there are multiple settings lines and the prefix for one line
  repeats the prefix of the most recent settings line, the repeated
  prefix can be omitted. So (reading '/' as a linebreak):
    'size w 600 / size h 600' can also be written as
    'size w 600 / h 600'.

* The internal field names use 1-character strings for some identifiers
  ('w' = width, 'h' = height, etc.), but these are not obvious to
  people. Therefore, if a user types the full word, the import routine
  will automatically translate it. This means that you can just type
  words that you would expect to work and they will:
  'margin left 10' = 'margin l 10', 'node height 20' = 'node h 20', etc.

* Boolean values for checkboxes can be written using
  'y' or 'yes' for true values and 'n' or 'no' for false, with any
  capitalization you wish.

All of this is in the spirit of Postel’s Law (a.k.a. the Robustness
Principle): "Be conservative in what you send, be liberal in what you
accept".

P.S. For the first time, the list of CSS color *names* can be used, not
just hex codes. (This currently only applies to these new Settings
lines, not in the lines declaring Flows or Nodes.)