Feature: 'Save my work' button produces a file with all inputs+settings
Added a prominent new 'Save' button; clicking it produces a plain text
file containing:
  * Headers with a timestamp and the SankeyMATIC URL
  * Your primary input lines
  * All settings in the diagram, using the new settings syntax

If you later copy & paste the entire file's contents into the Inputs
box, your diagram will be reproduced just as it was when saved
(excluding any manually moved nodes).

For now, the successful settings lines will still be present &
commented out (which looks untidy), but: if you then save the diagram
again, those lines will NOT be included in the file. This prevents
horrible cycles of continuing to add complete sets of inactive settings
lines to the saved output.