Feature: Pasted source code is tidied up when the diagram is rendered
When you paste in a diagram's saved source code, SankeyMATIC now looks
for all of the headers & settings lines it added at export time and
drops them from the pasted inputs.

The effect is that after your diagram is displayed, the Inputs box
will contain exactly what you had in it when you saved the diagram
and leave out all of the extra content and settings from the file.
Magic!

* This assumes that all of the settings lines were successfully
  applied, which should generally be a safe assumption. (The source
  file *could* have been manually updated to have an invalid value or
  name, in which case some settings lines will stick around.)