Fix: The layout algorithm was skipping the refinement iterations
The layouts being produced were only from the initial run of the
algorithm.

The iteration loop, which:
  * repositions nodes according to where they are attached, and
  * repeatedly re-centers the graph in order to provide the most room
    for nodes to spread apart evenly
...has not been running.

The bug was a very simple naming problem introduced in 4ea1fbd0 - the
field name was changed to internal_iterations in the index.html file
but not updated correspondingly in the main function. Fixed.

The same issue is also now fixed for the 'Reveal Shadows' checkbox.