Improved backward compatibility for diagrams saved before the new settings
Several new settings -- labelposition_scheme, labelvalue_position,
labels_magnify and labels_relativesize -- all have defaults which
produce very different visual output than the default behavior before
they existed.

Therefore, when importing a diagram’s source, we need to find a way to
preserve the appearance the user saw when they made the diagram.

To do so, we will prefix the old diagrams' inputs with values for those
settings which reproduce the pre-2024 behavior & appearance of labels.

It's ok to do this for *all* cases, in fact, because if someone is
importing a *post*-2023 diagram, later lines in the new diagram's
source will override the backfilled values provided by these lines
(following the "last one wins" rule).

---

One case not covered here is if the old diagram had set all label text
to ‘light’ or ‘bold’ (using labelname_weight). The Values in the new
diagram will always be normal-weight (due to the default value of
labelvalue_weight) and will *not* adopt the font-weight of the Name
portion if it is light or bold.
 * It would be possible to fix that with a bit more scripting by
   reading ahead in the source data, but I am expecting that this
   situation is 1) uncommon and 2) not difficult to notice & recover
   from if encountered.

Another case not covered is: What if an old diagram had '\n' as part of
a Node name? In the new version, that will become a multiline label
when it wasn't before! ...again, though, I expect that scenario will be
exceedingly unlikely and easy to fix if encountered.

---

P.S. The other new settings *not* mentioned above -- labels_linespacing
& labelposition_autoalign -- only matter when the backfilled settings
are set to some other value (i.e. labelposition_scheme = auto or
labelvalue_position = above/below). Since we won't have multiline
labels or auto-placed labels, those settings won't be used, and so they
don't need to be backfilled.