Feature: You can now hide labels while keeping the layout constant
There's a new checkbox, Labels > Arrange > "Temporarily Hide Labels".

Checking it keeps the current diagram’s spacing but displays no labels.

I have found this option useful when creating thumbnail versions of
diagrams for the site.

(This is distinct from the initial Show Names / Values checkboxes,
which can change the spacing on the sides of the diagram.)