Layout algorithm improvement. Overall symmetry should be increased.
Nodes which are not strictly required to be in lower stages are now
nudged to the right as far as they can be.

This mirrors the behavior of nodes toward the end of the diagram; each
is only as far away from its connected nodes as it strictly has to be.

One side effect:

- Now when the Layout checkboxes are used to force either Origins or
  Endpoints to the outer edges of the diagram, it is now easier for
  those outer nodes to arrive at suboptimal positions.

  However, this should be fixable in a straightforward way by
  reordering the source data rows as needed.

Also:

- Various refactoring/renaming to clarify assignNodesToStages' logic.