Harmonize incoming-flow calculation with remainder-consumption logic
The logic for filling in input gaps was not accounting for siblings.
Sample input demonstrating the issue:

  Wages [400] Budget
  Other [?] Budget
  Budget [500] Spent

"Other" should be calculated as 100, but it was not taking its sibling
"Wages" into account, so it was adopting 500 as its value (leading to a
new "Budget" node size of 900 rather than 500).

After this change, "Other" is calculated as 100.