Fix: Prevent errors when multiple consumers exist for a node.
When adding up the activity arriving at & leaving each parent node, we
now skip over any flows which are not valid numbers (which at this time
translates to: any other calculated flows).

This fix removes the need to skip over any flows between the *same*
source & target. It can now be valid input for there to be both a flow
with a defined amount and a calculated flow touching the same two
nodes.

Sample input which caused errors before and now does not:
  Wages [500] Budget
  Budget [*] Taxes
  Budget [*] Housing

Before this change:
* The sizes of both "Taxes" and "Housing" were computed as Not a Number
  (NaN).
* 'NaN' was passed in the SVG, so console errors would appear such as:
  <path> attribute d: Expected number, "M266.17045 NaNC387.25567 Na..."

After this change:
* "Taxes" = 500; "Housing" = 0. ("Taxes" consumed 100% of the remainder
  before "Housing"s calculation was run.)
* All flows are drawn with no console errors.