Fix: Auto-placed labels for 0-size nodes appear on the correct side
SankeyMATIC was using the SUM of flows out of a node to decide if that
node was an endpoint (and the sum of the inflows to decide if it was
an origin).

The sum of flows is the wrong measure to test in a world where we may
now calculate nodes of size 0 anywhere in the diagram.

Instead we now use the COUNT of flows out (or in) to determine if a
node's label should be automatically placed to one side or the other.