Update flow calculations to include nodes with -struck out names-
This feature is not widely known/used, but: you can add -dashes- around
any instance of a node's name to keep its label from being displayed.

Nodes with this unusual syntax were not yet being included in the flow
calculations (unless *every* mention of the node was -struck out-).

This change makes the calculations look at each node's "true" name so
that its amounts are incorporated whether extra dashes are present or
not.

A sample dataset which demonstrates the difference:
  Wages [100] Budget
  Budget [50] Transportation
  -Budget- [25] Fast food
  Budget [*] Housing

Before this change:
* "Housing" = 50 (100 - 50) because the remainder calculation skips
  the line for "Fast food"
* The Budget node is therefore shown (label-less) with an incorrect
  size of 125 (since flows of 50 + 25 + 50 are using it as a source)

  After:
  * "Housing" = 25 (100 - 50 - 25) and "Budget" = 100 (with no label)

Technical/refactoring notes:
* addNodeName (which returned nothing) has become setUpNode, which
  returns the node object itself
* getTrueNodeName is now parseNodeName
* getUniqueNode was no longer called & is deleted