Improve support for nodes & flows with size=0
Added some protections for the case where node sizes will be zero
(which can happen at any time with calculated flows):

1) Prevent division-by-0 instances by passing all candidate
calculations through a function which will replace any '0' denominator
with Number.MIN_VALUE. (This means that you basically get Infinity back
from such a calculation.)

2) In cases where the scale for a graph comes out as Infinity, just set
it to 1. (This reliably happens when every node in the graph is 0; see
below.)

3) While 0-size nodes have a display value of zero, their actual
internal value is now set to Number.MIN_VALUE, which means that they
can now participate in layout calculations in a useful way. So now, an
all-zeroes diagram like this:
  Wages [?] Budget
  Other [?] Budget
  Budget [*] Taxes
  Budget [*] Housing
  Budget [*] Food
is laid out in a readable fashion. (The Size and Scale sliders are
unpredictable with this kind of graph, but this is still an
improvement layout-wise.)

Finally, given the above, I have removed the ancient restriction that
you can't enter a flow with a literal size of [0]. It seems silly to
prevent someone entering one as data when calculated flows can now spit
out 0s left and right.