Fix: If the graph is reversed, flow calculations must also be reversed
When the Layout checkbox 'Reverse the graph' is checked, all flow
sources & targets are swapped, which means that the original
'*' and '?' symbols will now trigger the opposite calculation from what
was intended.

Here we add logic to make that work out correctly.

As far as examples of before/after, nearly every case of a calculated
flow will experience this issue.

The most trivial example:
  a [1] b
  b [*] c

Before this change: c = 0 when the graph is reversed.
After this change: c = 1 in the default rendering AND when reversed.