UI: Improve layout for narrow (mobile) windows
When viewing the Build interface on a phone, the layout has been quite
awkward: the diagram itself basically starts off-screen, to the right
of the viewport.

In the increasingly common use case of someone posting a Public Link to
a diagram, this has made for a fairly poor experience: anyone visiting
that link on a phone has to hunt around for the diagram for a bit to
find it.

This commit tries to improve the situation:

* The various Control sections are no longer embedded in an HTML table.
They are now somewhat independent <div>s that can occupy different
spots in a CSS Grid.
* Now, on a narrow screen:
  * All the controls plus the diagram are now shown in one vertical
    stack, with the inline margins dropped to 0.
  * The actual diagram is hoisted up to be one of the top items of the
    stack. This should make public diagrams far easier to find on
    mobile devices.
  * Below the diagram, the various Control sections are centered.
  * The layouts for some other very visible elements have also been
    improved (especially the Sample Diagrams section).

There are definitely more refinements that should be made here, but
this is a start.